feat: finish register route
