+++ backend/src/@types/db/index.ts
@@ -1,0 +1,7 @@
+export interface IAuthorTable {
+  id: string
+  username: string
+  password: string
+  refresh_token?: string
+  iat: string
+}
