+++ backend/src/@types/routes/authenticate/index.d.ts
@@ -1,0 +1,3 @@
+export interface IRegisterBody {
+  username: string
+}
